backend/requirements.txt
django==4.2.1
ipdb==0.13.13
flake8==6.0.0
django-environ==0.10.0
Pillow==9.4.0
celery[redis]==5.2.7
django-celery-results==2.5.0
django-extensions==3.2.1
django-celery-beat==2.5.0
pymemcache==4.0.0
django-lifecycle==1.0.0
django-rosetta==0.9.9
django-silk==5.0.3
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-factoryboy==2.5.1
psycopg2-binary
djangorestframework==3.14.0
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2
daphne==4.0.0